test(mcp-it): isolate TTL expiry test into its own IT class
